chore(test): consolidate object pool tests
